fix migrate bug

+++ database/migrations/2018_06_26_145219_update_short_url_records_3.php
@@ -32,7 +32,9 @@ class UpdateShortUrlRecords3 extends Migration
             $table->dropColumn('browser');
             $table->dropColumn('device');
             $table->dropColumn('platform');
-            $table->dropColumn('app');
+            if (Schema::table('short_url_records')->hasColumn('app')) {
+                $table->dropColumn('app');
+            }
         });
     }
 }
